Fix connection conflict (choose the lower one)
When two clients try to connect to one another only
the connection coming from the client with the lower ID stays
open, the other one is closed, this is implemented using mostly
the WrtcConnector.
There seem to be more connection problems but they're more subtle
and I don't want to debug those for now, we need to take this to wasm.